backend/requirements.txt
pymongo
python-dotenv
dnspython
flask